Move SnapshotPackagerService into its own module
